move callback intent service to separate class and add manifest declaration
